fix main styles
