+++ src/layout/Main/index.ts
@@ -1,0 +1,1 @@
+export { Main } from './Main';
